ディレクトリ構成変更 & モジュールの切り離しテスト
