+++ src/main/scala/app/demo/Demo.scala
@@ -1,5 +1,6 @@
-package com.github.imomushi8.systra.demo
+package app.demo
 
+import com.github.imomushi8.systra.chart._
 import com.github.imomushi8.systra.report._
 import com.github.imomushi8.systra.virtual.VirtualMarket
 import com.github.imomushi8.systra.core.entity._
@@ -10,18 +11,15 @@ import com.github.imomushi8.systra.core.util.Initial
 import cats.effect.IO
 import fs2._
 
-trait Demo extends Tradable[VirtualMarket]:
+object Demo extends Tradable[VirtualMarket]:
   def summerize[Memory: Initial](brains:       Seq[(String, Brain[VirtualMarket, Memory])],
                                  firstCapital: Price,
-                                 firstChart:   Chart): Seq[Pipe[IO, Chart, SummaryReport]] =
-    val initSub = SummarySubReport(0,0,0,0,0)
-    val initSummary = SummaryReport("", 0, 0, 0, OK, initSub, initSub, initSub, 0, 0, 0)
+                                 firstChart:   Chart): Seq[Pipe[IO, Chart, SummaryReport]] = 
     given initMarket: Initial[VirtualMarket] = VirtualMarket.initial(firstCapital, firstChart)
-    
     brains.map { case (name, brain) => (chartStream: Stream[IO, Chart]) => 
       chartStream
         .through(trade(brain))
-        .fold(initSummary) { 
+        .fold(Initial[SummaryReport]()) { 
           case (current, (_, _, TradeState(sampleSize, flag, buy, sell, capital, maxCapital, consWins, consLoses, _))) =>
             val buySub  = SummarySubReport(buy.winCount, buy.loseCount, buy.profit, buy.loss, buy.cost)
             val sellSub = SummarySubReport(sell.winCount, sell.loseCount, sell.profit, sell.loss, sell.cost)
@@ -31,3 +29,10 @@ trait Demo extends Tradable[VirtualMarket]:
             SummaryReport(name, sampleSize, firstCapital, capital, flag, buySub, sellSub, buySub+sellSub, maximalDrawDown, consecutiveWinCount, consecutiveLoseCount)
         }
     }
+
+  /** 開始メソッド */
+  def begin(ws: WebSocketStream): Ops = new Ops(ws)
+ 
+  class Ops(ws: WebSocketStream):
+    /** Demoの終了メソッド */
+    def end() = ws() >> { IO.println("Done DemoTrade") }
