+++ src/main/scala/com/github/imomushi8/systra/core/util/UtilConstant.scala
@@ -3,7 +3,6 @@ package com.github.imomushi8.systra.core.util
 import java.time.format.DateTimeFormatter
 
 val datetimeFormatter    = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.SSS")
-val csvDatetimeFormatter = DateTimeFormatter.ofPattern("yyyy.MM.dd HH:mm")
 
 inline val START  = "START  {}"
 inline val END    = "END    {}"
